backend/requirements.txt
fastapi
uvicorn
python-dotenv
joblib
pydantic
python-multipart
google-genai
pillow
numpy
pandas
scikit-learn